+++ banzai/context.py
@@ -1,4 +1,7 @@
+import json
+
 from collections import namedtuple
+from dramatiq import JSONEncoder
 
 
 def Context(args):
@@ -8,3 +11,8 @@ def Context(args):
         args_dict = args
     constructor = namedtuple('Context', args_dict.keys())
     return constructor(**args_dict)
+
+
+class ContextJSONEncoder(JSONEncoder):
+    def encode(self, data):
+        return json.dumps(data, separators=(",", ":"), default=str).encode("utf-8")