+++ banzai/settings.py
@@ -6,6 +6,7 @@ settings.py: Settings script for banzai.
     pipeline context.
 
 """
+import os
 import operator
 
 from banzai.utils.instrument_utils import InstrumentCriterion
@@ -61,7 +62,25 @@ CALIBRATION_STACKER_STAGE = {'BIAS': 'banzai.bias.BiasMaker',
 
 CALIBRATION_IMAGE_TYPES = ['BIAS', 'DARK', 'SKYFLAT']
 
+# Stack delays are expressed in milliseconds--namely, each is five minutes
+CALIBRATION_STACK_DELAYS = {'BIAS': 300000,
+                            'DARK': 300000,
+                            'SKYFLAT': 300000}
+
 SINISTRO_IMAGE_TYPES = ['BIAS', 'DARK', 'SKYFLAT', 'EXPOSE', 'STANDARD', 'TRAILED', 'EXPERIMENTAL']
+
+REDIS_QUEUE_NAMES = {'DEFAULT': 'default',
+                     'PROCESS_IMAGE': 'process_image',
+                     'SCHEDULE_STACK': 'schedule_stack'}
+
+SCHEDULE_STACKING_CRON_ENTRIES = {'coj': '30 06 * * *',
+                                  'cpt': '00 15 * * *',
+                                  'tfn': '30 17 * * *',
+                                  'lsc': '00 21 * * *',
+                                  'elp': '00 23 * * *',
+                                  'ogg': '00 03 * * *'}
+
+ASTROMETRY_SERVICE_URL = os.getenv('ASTROMETRY_SERVICE_URL', 'http://astrometry.lco.gtn/catalog/')
 
 
 def make_calibration_filename_function(calibration_type, attribute_filename_functions, telescope_filename_function):