+++ banzai/tests/test_end_to_end.py
@@ -5,6 +5,7 @@ import argparse
 import pytest
 import mock
 
+from banzai import settings
 from banzai.main import redis_broker
 from banzai.dbs import populate_calibration_table_with_bpms, create_db, get_session, CalibrationImage, get_timezone
 from banzai.utils import fits_utils, date_utils, file_utils
@@ -23,6 +24,18 @@ INSTRUMENTS = [os.path.join(site, os.path.basename(instrument_path)) for site in
 DAYS_OBS = [os.path.join(instrument, os.path.basename(dayobs_path)) for instrument in INSTRUMENTS
             for dayobs_path in glob(os.path.join(DATA_ROOT, instrument, '201*'))]
 
+ENCLOSURE_DICT = {
+    'fl11': 'domb',
+    'kb27': 'clma',
+    'fs02': 'clma',
+}
+
+TELESCOPE_DICT = {
+    'fl11': '1m0a',
+    'kb27': '0m4b',
+    'fs02': '2m0a',
+}
+
 
 def run_end_to_end_tests():
     parser = argparse.ArgumentParser()
@@ -32,7 +45,9 @@ def run_end_to_end_tests():
     args = parser.parse_args()
     os.chdir(args.code_path)
     command = 'python setup.py test -a "--durations=0 --junitxml={junit_file} -m {marker}"'
-    return os.system(command.format(junit_file=args.junit_file, marker=args.marker))
+
+    # Bitshift by 8 because Python encodes exit status in the leftmost 8 bits
+    return os.system(command.format(junit_file=args.junit_file, marker=args.marker)) >> 8
 
 
 def run_reduce_individual_frames(raw_filenames):
@@ -40,7 +55,7 @@ def run_reduce_individual_frames(raw_filenames):
         raw_path = os.path.join(DATA_ROOT, day_obs, 'raw')
         for filename in glob(os.path.join(raw_path, raw_filenames)):
             file_utils.post_to_archive_queue(filename, os.getenv('FITS_BROKER_URL'))
-    redis_broker.join('default')
+    redis_broker.join(settings.REDIS_QUEUE_NAMES['PROCESS_IMAGE'])
 
 
 def run_stack_calibrations(frame_type):
@@ -50,13 +65,16 @@ def run_stack_calibrations(frame_type):
         timezone = get_timezone(site, db_address=os.environ['DB_ADDRESS'])
         min_date, max_date = date_utils.get_min_and_max_dates(timezone, dayobs, return_string=True)
         command = 'banzai_stack_calibrations --raw-path {raw_path} --frame-type {frame_type} ' \
-                  '--site {site} --camera {camera} --min-date {min_date} --max-date {max_date} ' \
+                  '--site {site} --camera {camera} ' \
+                  '--enclosure {enclosure} --telescope {telescope} ' \
+                  '--min-date {min_date} --max-date {max_date} ' \
                   '--db-address={db_address} --ignore-schedulability --fpack'
-        command = command.format(raw_path=raw_path, frame_type=frame_type, site=site, camera=camera, min_date=min_date,
-                                 max_date=max_date, db_address=os.environ['DB_ADDRESS'])
+        command = command.format(raw_path=raw_path, frame_type=frame_type, site=site, camera=camera,
+                                 enclosure=ENCLOSURE_DICT[camera], telescope=TELESCOPE_DICT[camera],
+                                 min_date=min_date, max_date=max_date, db_address=os.environ['DB_ADDRESS'])
         logger.info('Running the following stacking command: {command}'.format(command=command))
         os.system(command)
-    redis_broker.join('default')
+    redis_broker.join(settings.REDIS_QUEUE_NAMES['SCHEDULE_STACK'])
 
 
 def mark_frames_as_good(raw_filenames):
@@ -92,7 +110,6 @@ def run_check_if_stacked_calibrations_were_created(raw_filenames, calibration_ty
     for day_obs in DAYS_OBS:
         created_stacked_calibrations += glob(os.path.join(DATA_ROOT, day_obs, 'processed',
                                                           '*' + calibration_type.lower() + '*.fits*'))
-    redis_broker.join('default')
     assert number_of_stacks_that_should_have_been_created > 0
     assert len(created_stacked_calibrations) == number_of_stacks_that_should_have_been_created
 
@@ -103,7 +120,6 @@ def run_check_if_stacked_calibrations_are_in_db(raw_filenames, calibration_type)
     calibrations_in_db = db_session.query(CalibrationImage).filter(CalibrationImage.type == calibration_type)
     calibrations_in_db = calibrations_in_db.filter(CalibrationImage.is_master).all()
     db_session.close()
-    redis_broker.join('default')
     assert number_of_stacks_that_should_have_been_created > 0
     assert len(calibrations_in_db) == number_of_stacks_that_should_have_been_created