+++ banzai/tests/test_need_to_process_image.py
@@ -24,7 +24,7 @@ def test_no_processing_if_previous_success(mock_can_process, mock_instrument, mo
     mock_instrument.return_value = FakeInstrument()
     mock_processed.return_value = FakeRealtimeImage(success=True, checksum=md5_hash1)
     mock_md5.return_value = md5_hash1
-    assert not need_to_process_image('test.fits', FakeContext())
+    assert not need_to_process_image('test.fits')
 
 
 @mock.patch('banzai.dbs.commit_processed_image')
@@ -38,7 +38,7 @@ def test_do_process_if_never_tried(mock_can_process, mock_instrument, mock_heade
     mock_instrument.return_value = FakeInstrument()
     mock_processed.return_value = FakeRealtimeImage(success=False, checksum=md5_hash1, tries=0)
     mock_md5.return_value = md5_hash1
-    assert need_to_process_image('test.fits', FakeContext())
+    assert need_to_process_image('test.fits')
 
 
 @mock.patch('banzai.dbs.commit_processed_image')
@@ -52,7 +52,7 @@ def test_do_process_if_tries_less_than_max(mock_can_process, mock_instrument, mo
     mock_instrument.return_value = FakeInstrument()
     mock_processed.return_value = FakeRealtimeImage(success=False, checksum=md5_hash1, tries=3)
     mock_md5.return_value = md5_hash1
-    assert need_to_process_image('test.fits', FakeContext(), max_tries=5)
+    assert need_to_process_image('test.fits', max_tries=5)
 
 
 @mock.patch('banzai.dbs.commit_processed_image')
@@ -67,7 +67,7 @@ def test_no_processing_if_tries_at_max(mock_can_process, mock_instrument, mock_h
     max_tries = 5
     mock_processed.return_value = FakeRealtimeImage(success=False, checksum=md5_hash1, tries=max_tries)
     mock_md5.return_value = md5_hash1
-    assert not need_to_process_image('test.fits', FakeContext(), max_tries=max_tries)
+    assert not need_to_process_image('test.fits', max_tries=max_tries)
 
 
 @mock.patch('banzai.dbs.commit_processed_image')
@@ -83,7 +83,7 @@ def test_do_process_if_new_checksum(mock_can_process, mock_instrument, mock_head
     mock_processed.return_value = image
     mock_md5.return_value = md5_hash2
     mock_instrument.return_value = FakeInstrument()
-    assert need_to_process_image('test.fits', True, db_address='')
+    assert need_to_process_image('test.fits', db_address='')
     assert not image.success
     assert image.tries == 0
     assert image.checksum == md5_hash2