+++ CHANGES.md
@@ -1,3 +1,27 @@
+0.21.0 (2019-03-25)
+-------------------
+- Significant refactor to the pipeline context and settings files. We have now
+  split settings that are static into the settings file and settings that can
+  change at runtime into the "runtime context". This is in preparation for 
+  running a task queue (e.g. dramatiq).
+
+0.20.1 (2019-03-14)
+-------------------
+- Fixed bug where gaia astrometry service solve failure was not handled properly
+
+0.20.0 (2019-03-11)
+-------------------
+- Configure pipeline to use GAIA-Astrometry.net service for WCS solves.
+
+0.19.4 (2019-03-05)
+-------------------
+- We no longer overwrite all SATURATE vales for the 0.4m telescopes. We now only use the
+  defaults if the value is missing or 0. We also set/check MAXLIN to the same value.
+- Require enclosure and telescope for instrument query
+- Avoid trying to reduce files that don't have fits extension in filename
+- Added entrypoint to update instruments table in database
+- Removed defunct entrypoint to reduce night by site
+
 0.19.3 (2019-02-13)
 -------------------
 - Fixed bug where master calibrations had wrong daydir in filename